Comment du QuadTree
Comment de divers trucs initilaiser mais jamais utiliser

+++ Sources/DLL/Arbre/Noeuds/NoeudTable.cpp
@@ -69,7 +69,7 @@ void NoeudTable::afficherConcret() const
 	glPushAttrib(GL_ALL_ATTRIB_BITS);
 	glTranslatef(TRANSLATE_X_NOEUD_TABLE, TRANSLATE_Y_NOEUD_TABLE, 0);
 	NoeudAbstrait::appliquerAfficher();
-	glStencilFunc(GL_ALWAYS, 0, -1);
+	glStencilFunc(GL_ALWAYS, 0, (GLuint)-1);
 	ControleurNuanceurs::obtenirInstance()->assignerEstTable(1);
 	liste_->dessiner();
 	ControleurNuanceurs::obtenirInstance()->assignerEstTable(0);
